Move ExecutorsTuple to executors/mod.rs
